move tests for new slice_plotter class structure
